+++ models/user.go
@@ -3,7 +3,7 @@ package models
 import "golang.org/x/crypto/bcrypt"
 
 type User struct {
-	Id       int64  `json:"id" gorm:"primary_key;auto_increment;not_null"`
+	Id       uint   `json:"id" gorm:"primary_key;auto_increment;not_null"`
 	Username string `json:"username"`
 	Email    string `json:"email" gorm:"unique"`
 	Password string `json:"password"`
