+++ service/profile_service.go
@@ -2,11 +2,12 @@ package service
 
 import (
 	"errors"
+
 	"github.com/teten-nugraha/go-dev-forum/models"
 	"github.com/teten-nugraha/go-dev-forum/repository"
 )
 
-func Profile(id int64) (models.User, error) {
+func Profile(id uint) (models.User, error) {
 
 	var userExist = repository.FindUserById(id)
 	if (models.User{}) == userExist {
